Завершена работа по замечаниям наставника. - Удалён файл consts.js за ненадобностью на данный момент. - Упразднены файлы моделей (points-model.js и offers.model.js). Соответствующие классы перенесены в main.js. - Свойство init теперь является методом. (Файл route-presenter.js). - Для формы редактирования добавлено новое свойство editable.
