--- src/model/offers-model.js
@@ -1,7 +1,0 @@
-import {getOffers} from '../mock/offers.js';
-
-export default class OffersModel {
-  offers = getOffers();
-
-  getOffers = () => this.offers;
-}
