--- src/model/points-model.js
@@ -1,9 +1,0 @@
-import {generatePoint} from '../mock/point.js';
-
-const POINTS_AMOUNT = 5;
-
-export default class PointsModel {
-  points = Array.from({length: POINTS_AMOUNT}, generatePoint);
-
-  getPoints = () => this.points;
-}
